backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlmodel==0.0.19
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
websockets==12.0
numpy==1.26.4
scikit-learn==1.5.0
torch==2.3.1
noise==1.2.2
aiofiles==23.2.1
httpx==0.27.0
python-dotenv==1.0.1
